+++ lib/src/parse_asset_id.dart
@@ -1,5 +1,11 @@
 const example = 'image-Tb9Ew8CXIwaY6R1kjMvI0uRR-2000x3000-jpg';
 
+/// Parse ref [String] into id, width, height and format attributes.
+/// example: 'image-Tb9Ew8CXIwaY6R1kjMvI0uRR-2000x3000-jpg'
+/// - id: Tb9Ew8CXIwaY6R1kjMvI0uRR
+/// - width: 2000
+/// - height: 3000
+/// - format: jpg
 Map<String, dynamic> parseAssetId(String ref) {
   final refSplit = ref.split('-');